Missing signature configurations crash application on HTTPS request.
Summary:
In order to the application be capable of access our Google Cloud Endpoints
implementation, it must be signed and the key SHA1 used to create an OAuth
client ID for Android apps at the application console.

WARNING: the keystore used to sign the application during the tests is not the
same used at Google Play, thus pay attention to which ANDROID_CLIENT_ID are you
using at your Cloud Enpoints backend.